Update Grafana dashboards datasource UID
The hardcoded 'prometheus' datasource UID in the Grafana dashboards for IOM jobs and spark operator detailed status has been replaced with a dynamic variable. This allows the dashboards to be flexibly connected with different Prometheus instances or configurations. The datasource is now identified by "${datasource}" which is a templating list option that is also introduced in this change.